don't depend on current location to locate code provided by user
when doing cd back and forth it could happen, under circumstances
I'm not fully aware of, that we would try to jump back to a
temporary-non-existent directory. This would trigger an error on
tests.

So I made the tests not to jump back and forth. Just jump to provided
location (nevermind what was the previous location or whether it still
exists right now.